% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{ALL_BUILDING_BLOCKS}
\alias{ALL_BUILDING_BLOCKS}
\title{ALL_BUILDING_BLOCKS}
\format{
An object of class \code{character} of length 9.
}
\usage{
ALL_BUILDING_BLOCKS
}
\description{
Allowed Building Blocks values
}
\keyword{internal}
