% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{ALL_EXCEL_AXES}
\alias{ALL_EXCEL_AXES}
\title{ALL_EXCEL_AXES}
\format{
An object of class \code{character} of length 6.
}
\usage{
ALL_EXCEL_AXES
}
\description{
Allowed Excel Axes
}
\keyword{internal}
