% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{ALL_EXCEL_DIMENSIONS}
\alias{ALL_EXCEL_DIMENSIONS}
\title{ALL_EXCEL_DIMENSIONS}
\format{
An object of class \code{character} of length 8.
}
\usage{
ALL_EXCEL_DIMENSIONS
}
\description{
Allowed Excel Dimensions Blocks values
}
\keyword{internal}
