% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{ALL_EXCEL_SHEETS}
\alias{ALL_EXCEL_SHEETS}
\title{ALL_EXCEL_SHEETS}
\format{
An object of class \code{character} of length 22.
}
\usage{
ALL_EXCEL_SHEETS
}
\description{
Required Excel sheets to be read by UI
}
\keyword{internal}
