% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-ui.R
\docType{data}
\name{AllBuildingBlocks}
\alias{AllBuildingBlocks}
\title{AllBuildingBlocks}
\format{
An object of class \code{character} of length 9.
}
\usage{
AllBuildingBlocks
}
\description{
Allowed Building Blocks values
}
\keyword{internal}
