% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-write-excel.R
\name{applyDataValidation}
\alias{applyDataValidation}
\title{applyDataValidation}
\usage{
applyDataValidation(
  value,
  data,
  sheetName,
  columnNames,
  excelObject,
  additionalRows = 100
)
}
\arguments{
\item{value}{Character string. Data validation listing as an Excel expression.
For instance `"'Lookup'!$L$2:$L$4"` to use values 
from `'Lookup'` Excel sheet between `L2` and `L4` cells.}

\item{data}{A data.frame previously added to the Excel sheet `sheetName`}

\item{sheetName}{Character string. Name of the sheet}

\item{columnNames}{Character string. Names of column variable to apply the dataValidation}

\item{excelObject}{An openxlsx workbook object}

\item{additionalRows}{Integer. Additional rows to apply dataValidation}
}
\value{
Invisibly returns `NULL`. 
**Side effect**: mutates the workbook by writing dataValidation listing
}
\description{
Write a dataValidation listing to cells in an Excel sheet
}
\keyword{Excel}
