% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-latest-released-paths.R
\name{buildGitHubRawURL}
\alias{buildGitHubRawURL}
\title{buildGitHubRawURL}
\usage{
buildGitHubRawURL(owner, repo, version, file)
}
\arguments{
\item{owner}{Repository owner}

\item{repo}{Repository name}

\item{version}{Version/tag}

\item{file}{File path within the repository}
}
\value{
GitHub raw URL
}
\description{
Build a GitHub raw URL from components
}
\keyword{internal}
