% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{EXCEL_MAPPING}
\alias{EXCEL_MAPPING}
\title{EXCEL_MAPPING}
\format{
An object of class \code{data.frame} with 98 rows and 4 columns.
}
\usage{
EXCEL_MAPPING
}
\description{
Dictionary mapping Excel to Qualification variables
}
\keyword{internal}
