% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{EXCEL_OPTIONS}
\alias{EXCEL_OPTIONS}
\title{EXCEL_OPTIONS}
\format{
An object of class \code{list} of length 4.
}
\usage{
EXCEL_OPTIONS
}
\description{
List of default Excel options
}
\keyword{datasets}
