% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-write-excel.R
\name{excelListingValue}
\alias{excelListingValue}
\title{excelListingValue}
\usage{
excelListingValue(data, columnName, sheetName)
}
\arguments{
\item{data}{A data.frame}

\item{columnName}{Character string. Name of column variable from which to get listing}

\item{sheetName}{Character string. Name of the sheet from which to get listing}
}
\value{
Character string corresponding to Excel expression
}
\description{
Create an Excel expression for listing values used in dataValidation
}
\keyword{Excel}
