% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{excelOption}
\alias{excelOption}
\title{excelOption}
\usage{
excelOption(
  allowedValues = NULL,
  nullAllowed = TRUE,
  naAllowed = FALSE,
  expectedLength = NULL
)
}
\arguments{
\item{allowedValues}{array of character strings allowed}

\item{nullAllowed}{logical for `NULL` values allowed}

\item{naAllowed}{logical for `NA` values allowed}

\item{expectedLength}{expected rows for data}
}
\description{
Because `ospsuite.utils::validateColumns()` has been deprecated in favor of `ospsuite.utils::validateIsOption()`, 
`excelOption()` uses `ospsuite.utils::characterOption()` with default appropriate for validating data read from Excel
}
\keyword{internal}
