% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-ui.R
\docType{data}
\name{excelOptions}
\alias{excelOptions}
\title{excelOptions}
\format{
An object of class \code{list} of length 3.
}
\usage{
excelOptions
}
\description{
List of default Excel options
}
\keyword{datasets}
