% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-excel.R
\name{excelToQualificationPlan}
\alias{excelToQualificationPlan}
\title{excelToQualificationPlan}
\usage{
excelToQualificationPlan(
  excelFile,
  qualificationPlan = "qualification_plan.json"
)
}
\arguments{
\item{excelFile}{Path of Excel file defining the qualification plan features}

\item{qualificationPlan}{Path of qualification plan exported file. Must have `.json` extension.}
}
\description{
Export Excel as qualification plan
}
