% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-ui.R
\name{excelUI}
\alias{excelUI}
\title{excelUI}
\usage{
excelUI(
  fileName = "qualification.xlsx",
  snapshotPaths,
  observedDataPaths,
  excelTemplate = NULL,
  qualificationPlan = NULL,
  toRemove = NULL
)
}
\arguments{
\item{fileName}{Name of the Excel file to be created}

\item{snapshotPaths}{List of project snapshots given by their URL or relative path}

\item{observedDataPaths}{List of observed data sets (which are not included into the projects)
given by their URL or relative path}

\item{excelTemplate}{Default excel template (only captions and lookup tables)}

\item{qualificationPlan}{Optional: existing qualification plan.
If `NULL`, at least 1 project must be included into the projects list above}

\item{toRemove}{Optional: projects to remove? Optional: observed data sets to remove?}
}
\description{
excelUI
}
