% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-qualification.R
\name{formatAxesSettings}
\alias{formatAxesSettings}
\title{formatAxesSettings}
\usage{
formatAxesSettings(axesSettings)
}
\arguments{
\item{axesSettings}{Content of a qualification plan}
}
\value{
data.frame with columns `ID` and `Path`
}
\description{
Get a data.frame of project IDs and Paths/URLs
}
\keyword{internal}
