% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-qualification.R
\name{formatGlobalAxesSettings}
\alias{formatGlobalAxesSettings}
\title{formatGlobalAxesSettings}
\usage{
formatGlobalAxesSettings(axesSettings, plotName)
}
\arguments{
\item{axesSettings}{Content of a qualification plan}

\item{plotName}{Name of plot for which axes settings are defined}
}
\value{
A data.frame with axes setting information
}
\description{
Format axes settings for use in qualification plans or reports.
}
\keyword{internal}
