% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-ui-qualification.R
\name{formatPlotSettings}
\alias{formatPlotSettings}
\title{formatPlotSettings}
\usage{
formatPlotSettings(plotSettings, fillEmpty = FALSE)
}
\arguments{
\item{plotSettings}{Content of a qualification plan}

\item{fillEmpty}{Logical. If `FALSE`, empty values are replaced by `NA`.
If `TRUE`, fill empty values with default qualification plan Plot Settings.}
}
\value{
A data.frame with plot settings information
}
\description{
Format plot settings into a standardized data.frame for further processing or reporting
}
\keyword{internal}
