% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-project.R
\name{getBBDataFromProjects}
\alias{getBBDataFromProjects}
\title{getBBDataFromProjects}
\usage{
getBBDataFromProjects(projectData, qualificationProjects = NULL)
}
\arguments{
\item{projectData}{A data.frame of project snapshots}

\item{qualificationProjects}{Optional: a character vector of project IDs from a qualification plan}
}
\value{
A data.frame with columns `Project`, `BB-Type`, `BB-Name`, `Parent-Project`
}
\description{
Get a data.frame of projects, type, name and parent project
}
\examples{

# Get the project data from a list of paths
snapshotPaths <- list(
  "Raltegravir" = file.path(
    "https://raw.githubusercontent.com",
    "Open-Systems-Pharmacology",
    "Raltegravir-Model",
    "v1.2",
    "Raltegravir-Model.json"
  ),
  "Atazanavir" = file.path(
    "https://raw.githubusercontent.com",
    "Open-Systems-Pharmacology",
    "Atazanavir-Model",
    "v1.2",
    "Atazanavir-Model.json"
  )
)
projectData <- getProjectsFromList(snapshotPaths)

# Get the simulations Observed Data from the projects
getBBDataFromProjects(projectData)

}
