% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-qualification.R
\name{getBBDataFromQualification}
\alias{getBBDataFromQualification}
\title{getBBDataFromQualification}
\usage{
getBBDataFromQualification(qualificationContent)
}
\arguments{
\item{qualificationContent}{Content of a qualification plan}
}
\value{
A data.frame with columns `Project`, `Simulation` and `Output`
}
\description{
Get a data.frame of projects, type, name and parent project
}
