% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-ui.R
\name{getBBFromSnapshot}
\alias{getBBFromSnapshot}
\title{getBBFromSnapshot}
\usage{
getBBFromSnapshot(snapshotPath, bbType = "Simulations")
}
\arguments{
\item{snapshotPath}{A path/URL to a project snapshot (JSON file)}

\item{bbType}{A building block type (e.g. `"Simulations"`, `"Individuals"`)}
}
\value{
A list of building block information
}
\description{
Get a list of building blocks from a project snapshot
}
\keyword{internal}
