% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-excel.R
\name{getCTPlotsFromExcel}
\alias{getCTPlotsFromExcel}
\title{getCTPlotsFromExcel}
\usage{
getCTPlotsFromExcel(data, mapping)
}
\arguments{
\item{data}{A data.frame of plot settings}

\item{mapping}{A data.frame mapping plot information to projects}
}
\value{
A list of ComparisonTimeProfile plots
}
\description{
Get qualification settings for ComparisonTimeProfile plots
}
\keyword{internal}
