% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-excel.R
\name{getDDIPlotsFromExcel}
\alias{getDDIPlotsFromExcel}
\title{getDDIPlotsFromExcel}
\usage{
getDDIPlotsFromExcel(data, mapping)
}
\arguments{
\item{data}{A data.frame of plot settings}

\item{mapping}{A data.frame mapping plot information to projects}
}
\value{
A list of DDIRatio plots
}
\description{
Get qualification settings for DDI Ratio plots
}
\keyword{internal}
