% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-excel.R
\name{getExcelSections}
\alias{getExcelSections}
\title{getExcelSections}
\usage{
getExcelSections(sectionData)
}
\arguments{
\item{sectionData}{A data.frame}
}
\value{
A nexted list of sections
}
\description{
Parse qualification plan sections
}
\keyword{internal}
