% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-excel.R
\name{getGOFPlotsFromExcel}
\alias{getGOFPlotsFromExcel}
\title{getGOFPlotsFromExcel}
\usage{
getGOFPlotsFromExcel(data, mapping)
}
\arguments{
\item{data}{A data.frame of plot settings}

\item{mapping}{A data.frame mapping plot information to projects}
}
\value{
A list of GOFMerged plots
}
\description{
Get qualification settings for GOFMerged plots
}
\keyword{internal}
