% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-excel.R
\name{getInputsFromExcel}
\alias{getInputsFromExcel}
\title{getInputsFromExcel}
\usage{
getInputsFromExcel(data)
}
\arguments{
\item{data}{A data.frame of Inputs settings}
}
\value{
A list of Inputs
}
\description{
Get qualification Inputs
}
\keyword{internal}
