% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-latest-released-paths.R
\name{getLatestReleaseTag}
\alias{getLatestReleaseTag}
\title{getLatestReleaseTag}
\usage{
getLatestReleaseTag(owner, repo, includePreReleases)
}
\arguments{
\item{owner}{Repository owner}

\item{repo}{Repository name}

\item{includePreReleases}{Logical indicating whether to include pre-releases}
}
\value{
Latest release tag name, or NULL if fetch fails
}
\description{
Get the latest release tag from a GitHub repository
}
\keyword{internal}
