% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-project.R
\name{getObsDataFromList}
\alias{getObsDataFromList}
\title{getObsDataFromList}
\usage{
getObsDataFromList(observedDataPaths)
}
\arguments{
\item{observedDataPaths}{List of observed data paths and types}
}
\value{
data.frame with columns `ID`, `Path` and `Type`
}
\description{
Get a data.frame of observed data IDs and Paths/URLs
}
\examples{
# Get the project data from a list of paths
observedDataPaths <- list(
  "A" = "ObsData/A.csv",
  "B" = "ObsData/B.csv",
  "A-B-DDI" = list(Path = "Projects/A-B-DDI.csv", Type = "DDIRatio")
)
getObsDataFromList(observedDataPaths)

}
