% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-qualification.R
\name{getObsDataFromQualification}
\alias{getObsDataFromQualification}
\title{getObsDataFromQualification}
\usage{
getObsDataFromQualification(qualificationContent)
}
\arguments{
\item{qualificationContent}{Content of a qualification plan}
}
\value{
data.frame with columns `ID`, `Path` and `Type`
}
\description{
Get a data.frame of observed data IDs and Paths/URLs
}
