% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-excel.R
\name{getPKPlotsFromExcel}
\alias{getPKPlotsFromExcel}
\title{getPKPlotsFromExcel}
\usage{
getPKPlotsFromExcel(data, mapping)
}
\arguments{
\item{data}{A data.frame of plot settings}

\item{mapping}{A data.frame mapping plot information to projects}
}
\value{
A list of PKRatio plots
}
\description{
Get qualification settings for PK Ratio plots
}
\keyword{internal}
