% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-excel.R
\name{getPlotSettingsFromExcel}
\alias{getPlotSettingsFromExcel}
\title{getPlotSettingsFromExcel}
\usage{
getPlotSettingsFromExcel(data)
}
\arguments{
\item{data}{A data.frame of plot settings}
}
\value{
A list of plot settings
}
\description{
Get qualification plot setting
}
\keyword{internal}
