% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-project.R
\name{getProjectsFromList}
\alias{getProjectsFromList}
\title{getProjectsFromList}
\usage{
getProjectsFromList(snapshotPaths)
}
\arguments{
\item{snapshotPaths}{List of project snapshots given by their URL or relative path}
}
\value{
data.frame with columns `ID` and `Path`
}
\description{
Get a data.frame of project IDs and Paths/URLs
}
\examples{

# Get the project data from a list of paths
snapshotPaths <- list(
  "Raltegravir" = file.path(
    "https://raw.githubusercontent.com",
    "Open-Systems-Pharmacology",
    "Raltegravir-Model",
    "v1.2",
    "Raltegravir-Model.json"
  ),
  "Atazanavir" = file.path(
    "https://raw.githubusercontent.com",
    "Open-Systems-Pharmacology",
    "Atazanavir-Model",
    "v1.2",
    "Atazanavir-Model.json"
  )
)
getProjectsFromList(snapshotPaths)

}
