% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-qualification.R
\name{getQualificationAllPlots}
\alias{getQualificationAllPlots}
\title{getQualificationAllPlots}
\usage{
getQualificationAllPlots(qualificationContent, simulationsOutputs)
}
\arguments{
\item{qualificationContent}{Content of a qualification plan}

\item{simulationsOutputs}{A data.frame of Project, Simulation and Output}
}
\value{
data.frame with columns
`Project`, `Simulation` and `Section Reference`
}
\description{
Extract a data.frame containing All Plots information
from the qualification plan content
}
\keyword{internal}
