% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-qualification.R
\name{getQualificationCTProfile}
\alias{getQualificationCTProfile}
\title{getQualificationCTProfile}
\usage{
getQualificationCTProfile(qualificationContent)
}
\arguments{
\item{qualificationContent}{Content of a qualification plan}
}
\value{
data.frame with columns
`Title`, `Section Reference`, `Simulation Duration`, `TimeUnit` and plot settings
}
\description{
Extract a data.frame containing comparison time (CT) profile information
from the qualification plan content
}
\keyword{internal}
