% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-qualification.R
\name{getQualificationDDIRatioMapping}
\alias{getQualificationDDIRatioMapping}
\title{getQualificationDDIRatioMapping}
\usage{
getQualificationDDIRatioMapping(qualificationContent)
}
\arguments{
\item{qualificationContent}{Content of a qualification plan}
}
\value{
A data.frame with the following columns
`Project`, `Simulation_Control`, `Simulation_Treatment`, `Output` and control/treatment settings
}
\description{
Extract a data.frame mapping DDI ratio identifiers to relevant DDI Ratio fields
}
\keyword{internal}
