% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-ui-qualification.R
\name{getQualificationGOFPlots}
\alias{getQualificationGOFPlots}
\title{getQualificationGOFPlots}
\usage{
getQualificationGOFPlots(qualificationContent)
}
\arguments{
\item{qualificationContent}{Content of a qualification plan}
}
\value{
data.frame with columns
`Title`, `Section Reference`, `Artifacts`, `PlotTypes`, `Groups` and plot settings
}
\description{
Extract a data.frame containing goodness of fit (GOF) plot information
from the qualification plan content
}
\keyword{internal}
