% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-qualification.R
\name{getQualificationInputs}
\alias{getQualificationInputs}
\title{getQualificationInputs}
\usage{
getQualificationInputs(qualificationContent)
}
\arguments{
\item{qualificationContent}{Content of a qualification plan}
}
\value{
data.frame with columns 'Project', 'BB-Type', 'BB-Name', and 'Section Reference'
}
\description{
Get a data.frame of qualification inputs with columns 'Project', 'BB-Type', 'BB-Name', and 'Section Reference'
}
\keyword{internal}
