% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-qualification.R
\name{getQualificationIntro}
\alias{getQualificationIntro}
\title{getQualificationIntro}
\usage{
getQualificationIntro(qualificationContent)
}
\arguments{
\item{qualificationContent}{Content of a qualification plan}
}
\value{
data.frame with `Path` column
}
\description{
Get a data.frame of project Path
}
\keyword{internal}
