% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-ui-qualification.R
\name{getQualificationPKRatio}
\alias{getQualificationPKRatio}
\title{getQualificationPKRatio}
\usage{
getQualificationPKRatio(qualificationContent)
}
\arguments{
\item{qualificationContent}{Content of a qualification plan}
}
\value{
A data.frame with following columns:
`Title`, `Section Reference`, `PK-Parameter`, `Artifacts` and Group settings
}
\description{
Extract PK ratio data from a qualification plan and return it as a data.frame with relevant columns
}
\keyword{internal}
