% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-ui-qualification.R
\name{getQualificationPKRatioMapping}
\alias{getQualificationPKRatioMapping}
\title{getQualificationPKRatioMapping}
\usage{
getQualificationPKRatioMapping(qualificationContent, simulationsOutputs)
}
\arguments{
\item{qualificationContent}{Content of a qualification plan}

\item{simulationsOutputs}{A data.frame of Project, Simulation and Output}
}
\value{
A data.frame with the following columns:
`Project`, `Simulation`, `Output`, `Plot Title`, `Group Title`, `Observed data`, and `ObservedDataRecordId`
}
\description{
Extract a data.frame mapping PK ratio identifiers to relevant PK Ratio fields
}
\keyword{internal}
