% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-ui-qualification.R
\name{getQualificationSimParam}
\alias{getQualificationSimParam}
\title{getQualificationSimParam}
\usage{
getQualificationSimParam(qualificationContent)
}
\arguments{
\item{qualificationContent}{Content of a qualification plan}
}
\value{
data.frame with columns 'Project', 'Parent Project', 'Parent Simulation', 'Path', and 'TargetSimulation'
}
\description{
Get a data.frame of qualification simulation parameters
with columns 'Project', 'Parent Project', 'Parent Simulation', 'Path', and 'TargetSimulation'
}
\keyword{internal}
