% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-qualification.R
\name{getQualificationStyles}
\alias{getQualificationStyles}
\title{getQualificationStyles}
\usage{
getQualificationStyles(
  data,
  commonProjects,
  qualificationProjects,
  projectVariable = "Project"
)
}
\arguments{
\item{data}{A data.frame with a project variable}

\item{commonProjects}{A vector of project IDs that are common between the data and the qualification plan}

\item{qualificationProjects}{A vector of project IDs that are new in the qualification plan}

\item{projectVariable}{Name of the project variable in the data.frame}
}
\description{
Get qualification styles for a data.frame based on project membership
}
\keyword{internal}
