% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-ui-qualification.R
\name{getSchemaVersion}
\alias{getSchemaVersion}
\title{getSchemaVersion}
\usage{
getSchemaVersion(qualificationContent)
}
\arguments{
\item{qualificationContent}{Content of a qualification plan}
}
\value{
A data.frame with `Qualification plan schema version` column
}
\description{
Extract Qualification plan schema version from a Qualification Plan
}
\keyword{internal}
