% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-project.R
\name{getSimulationsFromSnapshot}
\alias{getSimulationsFromSnapshot}
\title{getSimulationsFromSnapshot}
\usage{
getSimulationsFromSnapshot(snapshotPath)
}
\arguments{
\item{snapshotPath}{A path/URL to a project snapshot (JSON file)}
}
\value{
A list of simulations information
}
\description{
Get a list of simulations from a project snapshot
}
\keyword{internal}
