% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-project.R
\name{getSimulationsOutputsFromProjects}
\alias{getSimulationsOutputsFromProjects}
\title{getSimulationsOutputsFromProjects}
\usage{
getSimulationsOutputsFromProjects(projectData)
}
\arguments{
\item{projectData}{A data.frame of project snapshots}
}
\value{
A data.frame with columns `Project`, `Simulation` and `Output`
}
\description{
Get a data.frame of projects, simulations and outputs
}
\examples{

# Get the project data from a list of paths
snapshotPaths <- list(
  "Raltegravir" = file.path(
    "https://raw.githubusercontent.com",
    "Open-Systems-Pharmacology",
    "Raltegravir-Model",
    "v1.2",
    "Raltegravir-Model.json"
  ),
  "Atazanavir" = file.path(
    "https://raw.githubusercontent.com",
    "Open-Systems-Pharmacology",
    "Atazanavir-Model",
    "v1.2",
    "Atazanavir-Model.json"
  )
)
projectData <- getProjectsFromList(snapshotPaths)

# Get the simulations outputs from the projects
getSimulationsOutputsFromProjects(projectData)

}
