% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-excel.R
\name{groupAxesSettings}
\alias{groupAxesSettings}
\title{groupAxesSettings}
\usage{
groupAxesSettings(qualificationAxesSettings)
}
\arguments{
\item{qualificationAxesSettings}{A data.frame of axis settings}
}
\value{
A named list of axis settings
}
\description{
Group axis settings as a named list to export in qualification plan
}
\keyword{internal}
