% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{lookupData}
\alias{lookupData}
\title{lookupData}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33 rows and 18 columns.
}
\usage{
lookupData
}
\description{
data.frame of all lookup values
Allowing all definitions to be centralized in Qualification-Template
}
\keyword{internal}
