% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-excel.R
\name{mapDataToProject}
\alias{mapDataToProject}
\title{mapDataToProject}
\usage{
mapDataToProject(projectId, data, type)
}
\arguments{
\item{projectId}{A character identifier of project}

\item{data}{A data.frame of Building Block or Simulation Parameter data}

\item{type}{A character `"BB"` for Building Block and `"SimParam"` for Simulation Parameter data}
}
\value{
A formatted list or data.frame mapped to the Project
}
\description{
Map building block or simulation parameter data to a project
}
\keyword{internal}
