% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-excel.R
\name{mapToQualification}
\alias{mapToQualification}
\title{mapToQualification}
\usage{
mapToQualification(data, sheetName, qualificationPlanSelector = NA)
}
\arguments{
\item{data}{A data.frame or a list}

\item{sheetName}{Selected Excel sheet name}

\item{qualificationPlanSelector}{Qualification Plan selector for sub-fields}
}
\value{
A data.frame or a list
}
\description{
Map and rename qualification data.frame or list using dictionary
}
\keyword{internal}
