% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-ui.R
\name{mergeBBData}
\alias{mergeBBData}
\title{mergeBBData}
\usage{
mergeBBData(bbData, qualificationBBData)
}
\arguments{
\item{bbData}{A data.frame with columns `Project`, `BB-Type`, `BB-Name`, `Parent-Project`}

\item{qualificationBBData}{A data.frame with columns `Project`, `BB-Type`, `BB-Name`, `Parent-Project`}
}
\value{
A data.frame with columns `Project`, `BB-Type`, `BB-Name`, `Parent-Project`
}
\description{
Merge data.frames of Building Blocks data from snapshots and qualification
}
