% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-ui.R
\name{mergeObsData}
\alias{mergeObsData}
\title{mergeObsData}
\usage{
mergeObsData(obsData, qualificationObsData = NULL)
}
\arguments{
\item{obsData}{A data.frame with columns `Id`, `Path` and `Type`}

\item{qualificationObsData}{A data.frame with columns `Id`, `Path` and `Type`}
}
\value{
A data.frame with columns `Id`, `Path`, `Type` and `Status`
}
\description{
Merge data.frames of Observed data IDs and Paths/URLs from snapshots and qualification
}
