% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-ui.R
\name{mergeProjectData}
\alias{mergeProjectData}
\title{mergeProjectData}
\usage{
mergeProjectData(snapshotData, qualificationData = NULL)
}
\arguments{
\item{snapshotData}{A data.frame with columns `Id` and `Path`}

\item{qualificationData}{A data.frame with columns `Id` and `Path`}
}
\value{
A data.frame with columns `Id`, `Path` and `Status`
}
\description{
Merge data.frames of project IDs and Paths/URLs from snapshots and qualification
}
\examples{

# Qualification data
qualiData <- data.frame(Id = c("a", "b", "c"), Path = c("a", "b", "c"))

# Snapshot data
snapData <- data.frame(Id = c("c", "d"), Path = c("newC", "newD"))

# Merged data along with status
mergeProjectData(snapData, qualiData)

}
