% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-latest-released-paths.R
\name{parseGitHubURL}
\alias{parseGitHubURL}
\title{parseGitHubURL}
\usage{
parseGitHubURL(url)
}
\arguments{
\item{url}{GitHub raw URL}
}
\value{
A list with owner, repo, version, and file, or NULL if parsing fails
}
\description{
Parse a GitHub raw URL to extract owner, repo, version, and file
}
\keyword{internal}
