% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-qualification.R
\name{parseSectionsToDataFrame}
\alias{parseSectionsToDataFrame}
\title{parseSectionsToDataFrame}
\usage{
parseSectionsToDataFrame(
  sectionsIn,
  sectionsOut = data.frame(),
  parentSection = NA
)
}
\arguments{
\item{sectionsIn}{A Section list including Reference, Title, Content and Sections fields}

\item{sectionsOut}{A data.frame to accumulate the parsed sections}

\item{parentSection}{A string representing the parent section reference}
}
\value{
A data.frame
}
\description{
Parse qualification plan sections
}
\keyword{internal}
