% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{PLOT_SETTINGS}
\alias{PLOT_SETTINGS}
\title{PLOT_SETTINGS}
\format{
An object of class \code{list} of length 3.
}
\usage{
PLOT_SETTINGS
}
\description{
Default plot settings used by qualification plans
}
\keyword{internal}
