% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-latest-released-paths.R
\name{processObservedDataPaths}
\alias{processObservedDataPaths}
\title{processObservedDataPaths}
\usage{
processObservedDataPaths(
  qualificationContent,
  includePreReleases,
  returnUpdatedOnly,
  observedDataToIgnore
)
}
\arguments{
\item{qualificationContent}{Content of qualification plan}

\item{includePreReleases}{Logical indicating whether to include pre-releases}

\item{returnUpdatedOnly}{Logical indicating whether to return only updated paths}

\item{observedDataToIgnore}{Character vector of observed data IDs to ignore}
}
\value{
Character vector of observed data URLs with latest releases
}
\description{
Process observed data paths to get latest releases
}
\keyword{internal}
