% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-latest-released-paths.R
\name{processProjectPaths}
\alias{processProjectPaths}
\title{processProjectPaths}
\usage{
processProjectPaths(
  qualificationContent,
  includePreReleases,
  returnUpdatedOnly,
  projectsToIgnore
)
}
\arguments{
\item{qualificationContent}{Content of qualification plan}

\item{includePreReleases}{Logical indicating whether to include pre-releases}

\item{returnUpdatedOnly}{Logical indicating whether to return only updated paths}

\item{projectsToIgnore}{Character vector of project IDs to ignore}
}
\value{
Character vector of project URLs with latest releases
}
\description{
Process project paths to get latest releases
}
\keyword{internal}
