% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-ui.R
\name{styleColorMapping}
\alias{styleColorMapping}
\title{styleColorMapping}
\usage{
styleColorMapping(mapping, sheetName, excelObject, columnName = "Color")
}
\arguments{
\item{mapping}{A data.frame}

\item{sheetName}{Character string. Name of the sheet to write to}

\item{excelObject}{An openxlsx workbook object}

\item{columnName}{Character string. Name of the column where colors are defined}
}
\value{
Invisibly returns `NULL`. Side effect: mutates the workbook by writing data and freezing the header row.
}
\description{
Apply background color to mapping data.frame in excel object
}
\keyword{internal}
