% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-write-excel.R
\name{styleProjectStatus}
\alias{styleProjectStatus}
\title{styleProjectStatus}
\usage{
styleProjectStatus(projectIds, columns, statusMapping, sheetName, excelObject)
}
\arguments{
\item{projectIds}{A vector of project Ids}

\item{columns}{Indices of the columns to apply the styles to}

\item{statusMapping}{A data.frame mapping colors and status}

\item{sheetName}{Name of the sheet to write to}

\item{excelObject}{An openxlsx workbook object}
}
\description{
Apply color styles to cells in an Excel sheet depending on identified status
}
\keyword{internal}
