% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-qualification.R
\name{styleQualificationCells}
\alias{styleQualificationCells}
\title{styleQualificationCells}
\usage{
styleQualificationCells(
  qualificationStyles,
  columnIndices,
  sheetName,
  excelObject
)
}
\arguments{
\item{qualificationStyles}{A vector of qualification styles including the names `"New"` and `"Deleted"`}

\item{columnIndices}{Indices of the columns to apply the styles to}

\item{sheetName}{Name of the sheet to write to}

\item{excelObject}{An openxlsx workbook object}
}
\description{
Apply qualification styles to cells in an Excel sheet
}
\keyword{internal}
