% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-ui.R
\name{toExcelEditor}
\alias{toExcelEditor}
\title{toExcelEditor}
\usage{
toExcelEditor(
  fileName = "qualification.xlsx",
  snapshotPaths = NULL,
  observedDataPaths = NULL,
  excelTemplate = NULL,
  qualificationPlan = NULL
)
}
\arguments{
\item{fileName}{Character string. Name of the Excel file to be created.}

\item{snapshotPaths}{Named list of project snapshots given by their URL or relative path.}

\item{observedDataPaths}{Named list of observed data sets (which are not included into the projects)
given by their URL or relative path.}

\item{excelTemplate}{Character string. Path to an Excel template file (only captions and lookup tables).
If `NULL`, uses the default template from the package.}

\item{qualificationPlan}{Character string. Path, URL, or JSON string of an existing qualification plan.
If `NULL`, at least 1 project must be included in the snapshotPaths.}
}
\value{
Invisibly returns `NULL`. Side effect: creates an Excel file at the specified path.
}
\description{
toExcelEditor
}
