% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-latest-released-paths.R
\name{updatePathWithLatestRelease}
\alias{updatePathWithLatestRelease}
\title{updatePathWithLatestRelease}
\usage{
updatePathWithLatestRelease(path, includePreReleases, returnUpdatedOnly)
}
\arguments{
\item{path}{Original GitHub raw URL}

\item{includePreReleases}{Logical indicating whether to include pre-releases}

\item{returnUpdatedOnly}{Logical indicating whether to return only updated paths}
}
\value{
Updated URL with latest release tag, or NULL if no update needed
}
\description{
Update a GitHub raw URL with the latest release tag
}
\keyword{internal}
