% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-qualification.R
\name{formatAxesSettings}
\alias{formatAxesSettings}
\title{formatAxesSettings}
\usage{
formatAxesSettings(axesSettings)
}
\arguments{
\item{axesSettings}{Content of a qualification plan}
}
\value{
A data.frame with axes setting information
}
\description{
Format axes settings for use in qualification plans or reports.
}
\keyword{internal}
