% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-qualification.R
\name{formatPlotSettings}
\alias{formatPlotSettings}
\title{formatPlotSettings}
\usage{
formatPlotSettings(plotSettings)
}
\arguments{
\item{plotSettings}{Content of a qualification plan}
}
\value{
A data.frame with plot settings information
}
\description{
Format plot settings into a standardized data.frame for further processing or reporting
}
\keyword{internal}
