% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-excel.R
\name{getAllPlotsFromExcel}
\alias{getAllPlotsFromExcel}
\title{getAllPlotsFromExcel}
\usage{
getAllPlotsFromExcel(data)
}
\arguments{
\item{data}{A data.frame of plot settings}
}
\value{
A data.frame with columns 'Project', 'Simulation', and 'Section Reference', or NA if no data
}
\description{
Get qualification settings for AllPlots field
}
\keyword{internal}
