% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-project.R
\name{getBBFromSnapshot}
\alias{getBBFromSnapshot}
\title{getBBFromSnapshot}
\usage{
getBBFromSnapshot(snapshotPath, bbType = "Simulations")
}
\arguments{
\item{snapshotPath}{A path to a project snapshot (JSON file)}

\item{bbType}{A building block type (e.g. `"Simulations"`, `"Individuals"`)}
}
\value{
A list of simulations information
}
\description{
Get a list of simulations from a project snapshot
}
\keyword{internal}
