% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-excel.R
\name{getProjectsFromExcel}
\alias{getProjectsFromExcel}
\title{getProjectsFromExcel}
\usage{
getProjectsFromExcel(projectData, bbData)
}
\arguments{
\item{projectData}{A data.frame of project Id and Path}

\item{bbData}{A data.frame mapping Building Block to parent project}
}
\value{
A list of Project with their building blocks
}
\description{
Get qualification project if building blocks
}
\keyword{internal}
