% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-qualification.R
\name{getQualificationCTMapping}
\alias{getQualificationCTMapping}
\title{getQualificationCTMapping}
\usage{
getQualificationCTMapping(qualificationContent)
}
\arguments{
\item{qualificationContent}{Content of a qualification plan}
}
\value{
A data.frame with columns 
`Project`, `Simulation`, `Output` and relevant CT fields
}
\description{
Extract the comparison time (CT) mapping from a qualification plan,
returning a data.frame with mapping information for CT analysis.
}
\keyword{internal}
