% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-qualification.R
\name{getQualificationDDIRatio}
\alias{getQualificationDDIRatio}
\title{getQualificationDDIRatio}
\usage{
getQualificationDDIRatio(qualificationContent)
}
\arguments{
\item{qualificationContent}{Content of a qualification plan}
}
\value{
A data.frame with following columns:
`Title`, `Section Ref`, `PK-Parameter`, `Plot Type`, `Subunits`, `Artifacts` and legend settings
}
\description{
Extract DDI ratio data from a qualification plan and return it as a data.frame with relevant columns
}
\keyword{internal}
