% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-ui-qualification.R
\name{getQualificationGOFMapping}
\alias{getQualificationGOFMapping}
\title{getQualificationGOFMapping}
\usage{
getQualificationGOFMapping(
  qualificationContent,
  simulationsOutputs,
  simulationsObsData
)
}
\arguments{
\item{qualificationContent}{Content of a qualification plan}

\item{simulationsOutputs}{A data.frame of Project, Simulation and Output}

\item{simulationsObsData}{A data.frame of Project, Simulation and ObservedData}
}
\value{
A data.frame with columns
`Project`, `Simulation`, `Output` and relevant GOF fields
}
\description{
Extract the goodness of fit (GOF) mapping from a qualification plan,
returning a data.frame with mapping information for GOF analysis.
}
\keyword{internal}
