% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-ui-qualification.R
\name{getQualificationInputs}
\alias{getQualificationInputs}
\title{getQualificationInputs}
\usage{
getQualificationInputs(qualificationContent, bbData)
}
\arguments{
\item{qualificationContent}{Content of a qualification plan}

\item{bbData}{A data.frame of Project, BB-Type, BB-Name and Parent-Project}
}
\value{
data.frame with columns 'Project', 'BB-Type', 'BB-Name', and 'Section Reference'
}
\description{
Get a data.frame of qualification inputs with columns 'Project', 'BB-Type', 'BB-Name', and 'Section Reference'
}
\keyword{internal}
