% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-ui-qualification.R
\name{getQualificationIntro}
\alias{getQualificationIntro}
\title{getQualificationIntro}
\usage{
getQualificationIntro(qualificationContent)
}
\arguments{
\item{qualificationContent}{Content of a qualification plan}
}
\value{
data.frame with `Path` column
}
\description{
Extract intro paths from qualification content as a data.frame
}
\keyword{internal}
