% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-ui-qualification.R
\name{getQualificationPKRatioMapping}
\alias{getQualificationPKRatioMapping}
\title{getQualificationPKRatioMapping}
\usage{
getQualificationPKRatioMapping(qualificationContent)
}
\arguments{
\item{qualificationContent}{Content of a qualification plan}
}
\value{
A data.frame with the following columns
`Project`, `Simulation_Control`, `Simulation_Treatment`, `Output` and control/treatment settings
}
\description{
Extract a data.frame mapping DDI ratio identifiers to relevant PK Ratio fields
}
\keyword{internal}
