% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-qualification.R
\name{getQualificationSections}
\alias{getQualificationSections}
\title{getQualificationSections}
\usage{
getQualificationSections(qualificationContent)
}
\arguments{
\item{qualificationContent}{Content of a qualification plan}
}
\value{
A data.frame with `Section Reference`, `Title`, `Content` and `Parent Section` columns
}
\description{
Get a data.frame of qualification plan sections
}
\keyword{internal}
