% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-project.R
\name{getSimulationsFromSnapshot}
\alias{getSimulationsFromSnapshot}
\title{getSimulationsFromSnapshot}
\usage{
getSimulationsFromSnapshot(snapshotPath)
}
\arguments{
\item{snapshotPath}{A data.frame of project snapshots}
}
\value{
A list of simulations information
}
\description{
Get a list of simulations from a project snapshot
}
\keyword{internal}
