% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-excel.R
\name{parseSectionsToNestedList}
\alias{parseSectionsToNestedList}
\title{parseSectionsToNestedList}
\usage{
parseSectionsToNestedList(sectionsIn, sectionData)
}
\arguments{
\item{sectionsIn}{A data.frame row including Reference, Title, Content and Parent Section fields}

\item{sectionData}{A data.frame of all section information}
}
\value{
A nested list
}
\description{
Parse qualification plan sections
}
\keyword{internal}
