% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-ui.R
\name{writeDataToSheet}
\alias{writeDataToSheet}
\title{writeDataToSheet}
\usage{
writeDataToSheet(data, sheetName, excelObject)
}
\arguments{
\item{data}{A data.frame}

\item{sheetName}{Name of the sheet to write to}

\item{excelObject}{An openxlsx workbook object}
}
\description{
Write a data.frame to a specific sheet in an Excel file
}
\keyword{internal}
